Moved them out of the written package. This is the easiest way of preventing them from being analysed.
