+++ src/main/java/de/uni_stuttgart/it_rex/course/domain/ContentIndex.java
