+++ src/main/java/de/uni_stuttgart/it_rex/course/domain/Lecturer.java
@@ -1,4 +1,4 @@
-package de.uni_stuttgart.it_rex.course.domain.written;
+package de.uni_stuttgart.it_rex.course.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
