--- src/main/java/de/uni_stuttgart/it_rex/course/domain/written/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Custom domain entities.
- */
-package de.uni_stuttgart.it_rex.course.domain.written;
