+++ src/main/java/de/uni_stuttgart/it_rex/course/repository/written/ContentIndexRepository.java
@@ -1,13 +1,13 @@
 package de.uni_stuttgart.it_rex.course.repository.written;
 
-import de.uni_stuttgart.it_rex.course.domain.written.ContentIndex;
+import de.uni_stuttgart.it_rex.course.domain.ContentIndex;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.UUID;
 
 /**
- * Spring Data  repository for the ContentIndex entity.
+ * Spring Data repository for the ContentIndex entity.
  */
 @SuppressWarnings("unused")
 @Repository
