+++ src/main/java/de/uni_stuttgart/it_rex/course/service/written/ChapterService.java
@@ -1,6 +1,6 @@
 package de.uni_stuttgart.it_rex.course.service.written;
 
-import de.uni_stuttgart.it_rex.course.domain.written.Chapter;
+import de.uni_stuttgart.it_rex.course.domain.Chapter;
 import de.uni_stuttgart.it_rex.course.repository.written.ChapterRepository;
 import de.uni_stuttgart.it_rex.course.service.mapper.written.ChapterMapper;
 import org.slf4j.Logger;
